the to prototype html pages, which do show what we can make
the to prototype html pages, which do show what we can make. The game
based on bootstrap and the editor. The editor is closer to a final
data-model. As an HTML tool it should be base on bootstrap. The game
should use canvas instead. So the game vcan be standalone and extended
withexternal js-libraries.